# Improve streebog performance