… streebog/src/consts.rs
… streebog/src/table.rs