Before:

```
     Running target/release/deps/streebog256-6ccc72afa032ebff

running 6 tests
test bench_16  ... bench:         289 ns/iter (+/- 0) = 55 MB/s
test bench_1k  ... bench:      18,017 ns/iter (+/- 134) = 56 MB/s
test bench_256 ... bench:       4,507 ns/iter (+/- 11) = 56 MB/s
test bench_64  ... bench:       1,131 ns/iter (+/- 4) = 56 MB/s
test bench_64k ... bench:   1,151,937 ns/iter (+/- 4,830) = 56 MB/s
test bench_8k  ... bench:     143,984 ns/iter (+/- 240) = 56 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured

     Running target/release/deps/streebog512-392d31e8a448df2d

running 6 tests
test bench_16  ... bench:         284 ns/iter (+/- 0) = 56 MB/s
test bench_1k  ... bench:      17,927 ns/iter (+/- 244) = 57 MB/s
test bench_256 ... bench:       4,488 ns/iter (+/- 250) = 57 MB/s
test bench_64  ... bench:       1,125 ns/iter (+/- 13) = 56 MB/s
test bench_64k ... bench:   1,176,670 ns/iter (+/- 5,938) = 55 MB/s
test bench_8k  ... bench:     147,063 ns/iter (+/- 275) = 55 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured
```


After:

```
     Running target/release/deps/streebog256-6ccc72afa032ebff

running 6 tests
test bench_16  ... bench:         216 ns/iter (+/- 0) = 74 MB/s
test bench_1k  ... bench:      13,478 ns/iter (+/- 6) = 75 MB/s
test bench_256 ... bench:       3,394 ns/iter (+/- 7) = 75 MB/s
test bench_64  ... bench:         853 ns/iter (+/- 5) = 75 MB/s
test bench_64k ... bench:     867,464 ns/iter (+/- 483) = 75 MB/s
test bench_8k  ... bench:     108,429 ns/iter (+/- 100) = 75 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured

     Running target/release/deps/streebog512-392d31e8a448df2d

running 6 tests
test bench_16  ... bench:         215 ns/iter (+/- 3) = 74 MB/s
test bench_1k  ... bench:      13,496 ns/iter (+/- 7,361) = 75 MB/s
test bench_256 ... bench:       3,375 ns/iter (+/- 11) = 75 MB/s
test bench_64  ... bench:         848 ns/iter (+/- 358) = 75 MB/s
test bench_64k ... bench:     862,798 ns/iter (+/- 2,509) = 75 MB/s
test bench_8k  ... bench:     107,882 ns/iter (+/- 1,921) = 75 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured
```